- Move "configure.hpp" and "WinMainBoot.hpp" into subdirectory. - Fix window procedure's name and not to use `WMB_WINDOWPROC` macro. - Add `WMB_USE_RESOURCES` macro. - Delete blank "resource.h".
