+++ FrameWindow/FrameWindow.vcxproj
@@ -155,8 +155,8 @@
     <ClCompile Include="winmain\winmain.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="configure.hpp" />
-    <ClInclude Include="WinMainBoot.hpp" />
+    <ClInclude Include="app\configure.hpp" />
+    <ClInclude Include="winmain\WinMainBoot.hpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
