+++ FrameWindow/FrameWindow.vcxproj.filters
@@ -17,7 +17,11 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="configure.hpp" />
-    <ClInclude Include="WinMainBoot.hpp" />
+    <ClInclude Include="winmain\WinMainBoot.hpp">
+      <Filter>winmain</Filter>
+    </ClInclude>
+    <ClInclude Include="app\configure.hpp">
+      <Filter>app</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
