--- FrameWindow/resources/resource.h
@@ -1,1 +1,0 @@
-﻿// Blank file
